moved top-level rendering logic to a separate component
