+++ src/components/App.js
@@ -2,24 +2,31 @@ import '../styles/app.css';
 import React, { useState, useEffect } from 'react';
 import { getWeather, parseWeather } from '../api/weather';
 import Search from './Search';
-import Message from './generic/Message';
-import Current from './Current';
-import Forecast from './Forecast';
+import Content from './Content';
+// import Footer from './Footer';
 
 const App = () => {
 	const [searchTerm, setSearchTerm] = useState('');
 	const [notFound, setNotFound] = useState(false);
 	const [location, setLocation] = useState(undefined);
-	const [currentWeather, setCurrentWeather] = useState(undefined);
+	const [current, setCurrent] = useState(undefined);
 	const [forecast, setForecast] = useState(undefined);
 	const [units, setUnits] = useState('celsius');
+
+	// useEffect(() => {
+	// 	window.navigator.geolocation.getCurrentPosition(
+	// 		({ coords }) =>
+	// 			setLocation({ lat: coords.latitude, lon: coords.longitude }),
+	// 		err => console.log('Geolocation not available or user denied.')
+	// 	);
+	// }, []);
 
 	useEffect(async () => {
 		setNotFound(false);
 		if (searchTerm) {
 			try {
 				const { data } = await getWeather.currentByTerm(searchTerm);
-				setCurrentWeather(parseWeather.current(data));
+				setCurrent(parseWeather.current(data));
 				setLocation(parseWeather.coords(data));
 			} catch {
 				setNotFound(true);
@@ -28,61 +35,37 @@ const App = () => {
 	}, [searchTerm]);
 
 	useEffect(async () => {
+		// if (!current && location) {
+		// 	const { data } = await getWeather.currentByGeo(location);
+		// 	setCurrent(parseWeather.current(data));
+		// }
 		if (location) {
 			const { data } = await getWeather.forecast(location);
 			setForecast(parseWeather.forecast(data));
 		}
-	}, [location]);
+	}, [location /*, current*/]);
 
 	const switchUnits = () => {
 		units === 'celsius' ? setUnits('fahrenheit') : setUnits('celsius');
 	};
 
-	const renderError = () => {
-		if (notFound) {
-			return (
-				<Message
-					className="error message"
-					message={`! Sorry, "${searchTerm}" wasn't found. Try search for a bigger settlement nearby. !`}
-				/>
-			);
-		}
-	};
-
-	const renderCurrent = () => {
-		if (currentWeather) {
-			return (
-				<Current
-					data={currentWeather}
-					units={units}
-					switchUnits={switchUnits}
-				/>
-			);
-		} else {
-			return (
-				<Message
-					className="message"
-					message="To find out how the weather would be like, do manual search or enable GPS localization..."
-				/>
-			);
-		}
-	};
-
-	const renderForecast = () => {
-		if (forecast) {
-			return (
-				<Forecast data={forecast} units={units} switchUnits={switchUnits} />
-			);
-		}
+	const props = {
+		searchTerm,
+		notFound,
+		current,
+		forecast,
+		units,
+		switchUnits
 	};
 
 	return (
-		<div className="container">
-			<Search setSearchTerm={setSearchTerm} />
-			{renderError()}
-			{renderCurrent()}
-			{renderForecast()}
-		</div>
+		<React.Fragment>
+			<div className="container">
+				<Search setSearchTerm={setSearchTerm} />
+				<Content {...props} />
+			</div>
+			{/* <Footer /> */}
+		</React.Fragment>
 	);
 };
